Increase blackbox-exporter memory limit to 100Mi
We've seen blackbox-exporter OOM kills with the previous memory
limit of 40Mi on some clusters. Since increasing the memory limit
doesn't change scheduling behavior, we go ahead and increase the default
limit in the component defaults rather than adjusting it on individual
clusters.

When the blackbox-exporter is OOM-killed, it's possible that we get
spurious alerts for the ApiServer canary SLO, if one probe is missed due
to the pod restart.