Guard against division by zero in storage SLO
On some clusters which have very little workload churn, there may be
time windows in which there are 0 storage operations. The current
`total_query` will evaluate to 0 in such a time window, leading to a
division by zero for the SLO itself, which causes spurious alerts.

This commit adjusts the `total_query` to guard against this case by
wrapping the previous expression with `(expr > 0) or on() vector(1)`
which evaluates to `expr` if the expression results in > 0 and to 1
otherwise.

This removes the possibility of the SLO evaluating to NaN due to
dividing by zero. Note that we can safely use `vector(1)` in the guard,
as there can't be any failed storage operations if there's no storage
operations at all. Therefore, if we divide by `vector(1)`, the
expression always reduces to 0/1.