Adjust storage SLO error query to always result in a value
When there's no failing operations in a time window, it's possible that
the status="fail-unknown" storage_operation_duration_seconds_count time
series doesn't exist at all, which results in an empty vector for the
current error query.

This commit extends the error query as `expr or on() vector(0)` to
ensure we always return a value. The updated query returns 0, if the
fail-unknown time series doesn't exist for a time window.