Allow root-less pings on non-OpenShift
As mentioned in
https://github.com/appuio/network-canary#unprivileged-icmp, the
network-canary uses unprivileged pings.

After searching for a while, I found that we need to set the mentioned
sysctl on the pod for it to actually make a difference (it's enabled on
the node).

From what I can see, this is not required on OpenShift 4 for some
reason, so I'm just setting it for non-OpenShift.