Stop canary pods stopping downscaling. (#86)
The canary pods are seen as having no controller (deployment, ...) and thus block the removal of nodes. The problem is accentuated because they are usually scheduled on the least used node - which is also the node that would be downscaled.

As a follow-up we might want to update the controller scheduling those pods to be seen as a controller similarly to a deployment.